+++ src/index.ts
@@ -6,6 +6,7 @@ import { dialogActions } from './store/dialogSlice';
 import { appActions } from './store/appSlice';
 import { useAuthTokenCheck } from "./hooks/useAuthTokenCheck";
 import PreventAccidentalNavigationPrompt from './components/misc/PreventAccidentalNavigationPrompt';
+import InternalNavigator from './components/misc/InternalNavigator';
 
 const coreReduxActions = {
   appActions,
@@ -20,4 +21,5 @@ export {
   coreReduxActions,
   useAuthTokenCheck,
   PreventAccidentalNavigationPrompt,
+  InternalNavigator,
 }